Install tools for all installed versions of VS
