+++ SolutionInfo.proj
@@ -5,7 +5,7 @@
     <Copyright>© 2025 Shawn Marlin</Copyright>
     <Product>StatePipes</Product> 
     <ProductName>StatePipes</ProductName>
-    <VersionPrefix>2.0.8</VersionPrefix>
+    <VersionPrefix>2.0.9</VersionPrefix>
     <VersionSuffix></VersionSuffix>
     <SourceRevisionId></SourceRevisionId>
     <RepositoryType>git</RepositoryType>
