+++ StatePipes.ServiceCreatorToolSetup/SetupEnvironment.cs
@@ -1,4 +1,5 @@
-﻿using System.Xml.Linq;
+﻿using Microsoft.VisualStudio.Setup.Configuration;
+using System.Xml.Linq;
 
 namespace StatePipes.ServiceCreatorToolSetup
 {
@@ -6,27 +7,49 @@ namespace StatePipes.ServiceCreatorToolSetup
     {
         private const string statePipesLocalNugetsEnvironmentVariableName = "StatePipesLocalNugets";
         private const string statePipesPrivateNugets = "StatePipes Private Nugets";
-
+        private static void SetupVSSettings(System.Diagnostics.Process vsProcess)
+        {
+            System.Threading.Thread.Sleep(10000);
+            var dte = ExternalDTE.GetDTE2(vsProcess.Id);
+            if (dte == null) throw new Exception($"Couldn't find process {vsProcess.ProcessName} id {vsProcess.Id}");
+            string resourceFileName = $"{typeof(ImportSettings).Namespace}.Resources.StatePipesExternalToolsSettings.vssettings";
+            ImportSettings.ImportSettingsFromResource(dte, resourceFileName);
+        }
+        private static bool IsPreRelease(ISetupInstance setupInstance)
+        {
+            ISetupInstanceCatalog setupInstanceCatalog = (ISetupInstanceCatalog)setupInstance;
+            return setupInstanceCatalog.IsPrerelease();
+        }
+        private static void SetupToolsForInstances(IEnumSetupInstances enumerator, string tempTextFileName)
+        {
+            System.Diagnostics.Process? vsProcess = null;
+            int count = 1;
+            while (count == 1)
+            {
+                ISetupInstance[] setupInstances = new ISetupInstance[1];
+                enumerator.Next(1, setupInstances, out count);
+                if (count != 1 || setupInstances[0] == null || IsPreRelease(setupInstances[0])) continue;
+                string installationPath = setupInstances[0].GetInstallationPath();
+                string executablePath = Path.Combine(installationPath, @"Common7\IDE\devenv.exe");
+                vsProcess = System.Diagnostics.Process.Start(executablePath, $"{tempTextFileName} /nosplash");
+                System.Threading.Thread.Sleep(10000);
+                SetupVSSettings(vsProcess);
+                vsProcess?.Kill();
+            }
+        }
         private static void SetupTools()
         {
             string tempFileName = Path.GetTempFileName();
             string tempTextFileName = Path.ChangeExtension(tempFileName, ".txt");
             File.WriteAllText(tempTextFileName, "Do Not Touch. This instance of visual studio will close automatically when done adding external tools.");
-            var vsProcess = VisualStudioLauncher.LaunchSolution(tempTextFileName);
-            Thread.Sleep(10000);
             try
             {
-                var dte = ExternalDTE.GetDTE2(vsProcess.Id);
-                if (dte == null) throw new Exception($"Couldn't find process {vsProcess.ProcessName} id {vsProcess.Id}");
-                string resourceFileName = $"{typeof(ImportSettings).Namespace}.Resources.StatePipesExternalToolsSettings.vssettings";
-                ImportSettings.ImportSettingsFromResource(dte, resourceFileName);
+                ISetupConfiguration setupConfiguration = new SetupConfiguration();
+                IEnumSetupInstances enumerator = setupConfiguration.EnumInstances();
+                SetupToolsForInstances(enumerator, tempTextFileName);
             }
             catch { throw; }
-            finally
-            {
-                vsProcess.Kill();
-                if (File.Exists(tempTextFileName)) File.Delete(tempTextFileName);
-            }
+            finally { if (File.Exists(tempTextFileName)) File.Delete(tempTextFileName); }
         }
         private static void SetupNugets()
         {
